move defineWordEnding to utils.js, renamed util to utils
